Update test assertions for numerator and denominator types
Changed the assertions in `test_intelligence.py` to validate that `numerator` and `denominator` are floats instead of integers. This aligns the test expectations with changes in the data model or input format.

+++ tests/integration/test_intelligence.py
@@ -161,8 +161,8 @@ class TestIntelligenceEndpoints(BaseTestCase):
                     assert isinstance(event.symbol, str)
                     assert isinstance(event.split_event_date, date)
                     assert isinstance(event.label, str)
-                    assert isinstance(event.numerator, int)
-                    assert isinstance(event.denominator, int)
+                    assert isinstance(event.numerator, float)
+                    assert isinstance(event.denominator, float)
 
     def test_get_ipo_calendar(self, fmp_client: FMPDataClient, vcr_instance):
         """Test getting IPO calendar"""
